make the protector config only available from inside the protector
