+++ src/Contracts/ProtectorConfigContract.php
@@ -2,7 +2,6 @@
 
 namespace Cybex\Protector\Contracts;
 
-use Cybex\Protector\Exceptions\InvalidConnectionException;
 use Cybex\Protector\Exceptions\UnsupportedDatabaseException;
 use Illuminate\Contracts\Filesystem\Filesystem;
 use Illuminate\Database\Schema\SchemaState;
@@ -10,6 +9,25 @@ use Illuminate\Support\Collection;
 
 interface ProtectorConfigContract
 {
+    public function __construct(
+        ?bool $createDb = null,
+        ?bool $dropDb = null,
+        ?bool $tablespaces = null,
+        ?bool $dumpComments = null,
+        ?bool $dumpCharsets = null,
+        ?bool $dumpData = null,
+        ?bool $removeAutoIncrementingState = null,
+        ?string $connectionName = null,
+        ?string $dumpEndpointUrl = null,
+        ?string $authToken = null,
+        ?string $basicAuth = null,
+        ?string $privateKey = null,
+        ?string $maxPacketLength = null,
+        ?int $chunkSize = null,
+        ?int $httpTimeout = null,
+        ?array $metadataProviders = null
+    );
+
     /**
      * Returns the config value for the baseDirectory key.
      */
@@ -26,21 +44,11 @@ interface ProtectorConfigContract
     public function getConnectionConfig(): array|false;
 
     /**
-     * Returns the database config for the given connection.
-     */
-    public function getDatabaseConfig(): array|false;
-
-    /**
      * Returns the database name specified in the connectionConfig array.
      */
     public function getDatabaseName(): string;
 
     public function getConnectionName(): string;
-
-    /**
-     * @throws InvalidConnectionException
-     */
-    public function setConnectionName(?string $connectionName): static;
 
     public function getAuthToken(): string;
 
@@ -49,25 +57,10 @@ interface ProtectorConfigContract
      */
     public function getAuthTokenKeyName(): string;
 
-    public function setAuthToken(string $authToken): static;
-
-    public function withDefaultAuthToken(): static;
-
     /**
      * Basic Auth may only be used without Laravel Sanctum.
      */
     public function getBasicAuthCredentials(): ?string;
-
-    /**
-     * Basic Auth may only be used without Laravel Sanctum.
-     * The value format should be "<username>:<password>".
-     */
-    public function setBasicAuthCredentials(string $credentials): static;
-
-    /**
-     * Basic Auth may only be used without Laravel Sanctum.
-     */
-    public function withDefaultBasicAuthCredentials(): static;
 
     public function getPrivateKey(): string;
 
@@ -76,20 +69,12 @@ interface ProtectorConfigContract
      */
     public function getPrivateKeyName(): string;
 
-    public function setPrivateKey(string $privateKey): static;
-
-    public function withDefaultPrivateKey(): static;
-
     public function getDumpEndpointUrl(): string;
 
     /**
      * Gets the name of the .env key for the Protector dump endpoint URL.
      */
     public function getDumpEndpointUrlKeyName(): string;
-
-    public function setDumpEndpointUrl(string $dumpEndpointUrl): static;
-
-    public function withDefaultDumpEndpointUrl(): static;
 
     /**
      * Returns the maximum packet length specified in the config.
@@ -99,32 +84,9 @@ interface ProtectorConfigContract
      */
     public function getMaxPacketLength(): string;
 
-    /**
-     * The option 'max_allowed_packet' sets an upper limit on the size of any single message between the MySQL server and clients.
-     * This has to be set up on the client (here) and the MySQL server.
-     * This is not applicable to PostgreSQL.
-     */
-    public function setMaxPacketLength(string $maxPacketLength): static;
-
-    /**
-     * Uses the max packet length specified in the config file.
-     * The option 'max_allowed_packet' sets an upper limit on the size of any single message between the MySQL server and clients.
-     * This has to be set up on the client (here) and the MySQL server.
-     * This is not applicable to PostgreSQL.
-     */
-    public function withDefaultMaxPacketLength(): static;
-
     public function getChunkSize(): int;
 
-    public function setChunkSize(int $chunkSize): static;
-
-    public function withDefaultChunkSize(): static;
-
     public function getHttpTimeout(): int;
-
-    public function setHttpTimeout(int $httpTimeout): static;
-
-    public function withDefaultHttpTimeout(): static;
 
     /**
      * The metadata provider classes can be configured on the protector instance, else we return the config default.
@@ -133,61 +95,19 @@ interface ProtectorConfigContract
      */
     public function getMetadataProviders(): Collection;
 
-    public function setMetadataProviders(array $metadataProviders): static;
-
-    public function withDefaultMetadataProviders(): static;
-
     public function shouldRemoveAutoIncrementingState(): bool;
-
-    public function withAutoIncrementingState(): static;
-
-    public function withoutAutoIncrementingState(): static;
 
     public function shouldDumpCharsets(): bool;
 
-    public function withCharsets(): static;
-
-    public function withoutCharsets(): static;
-
     public function shouldDumpComments(): bool;
-
-    public function withComments(): static;
-
-    public function withoutComments(): static;
 
     public function shouldCreateDb(): bool;
 
-    public function withCreateDb(): static;
-
-    public function withoutCreateDb(): static;
-
     public function shouldDumpData(): bool;
-
-    public function withData(): static;
-
-    public function withoutData(): static;
 
     public function shouldDropDb(): bool;
 
-    /**
-     * Defines that existing databases should be dropped before importing a dump.
-     * Only works if used together with the $createDb option.
-     * (PostgreSQL only, controls the --clean flag)
-     */
-    public function withDropDb(): static;
-
-    /**
-     * Defines that existing databases should not be dropped before importing a dump.
-     * Only works if used together with the $createDb option
-     * (PostgreSQL only, controls the --clean flag)
-     */
-    public function withoutDropDb(): static;
-
     public function shouldUseTablespaces(): bool;
-
-    public function withTablespaces(): static;
-
-    public function withoutTablespaces(): static;
 
     public function shouldEncrypt(): bool;
 
@@ -196,11 +116,5 @@ interface ProtectorConfigContract
      */
     public function getProxyForSchemaState(): SchemaState;
 
-    /**
-     * Gets the current schema state parameters.
-     * These may change between calls, as the protector could be reconfigured to use a different connection and thus a different schema state proxy.
-     *
-     * @return array
-     */
     public function getSchemaStateParameters(): array;
 }
